Moved data to separated dirs
