+++ client/lib/models/session_status.dart
@@ -69,41 +69,3 @@ int compareSessionEntries(
   }
   return bTime.compareTo(aTime);
 }
-
-// Date/time formatting
-
-const _weekdays = ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'];
-const _months = [
-  'Jan',
-  'Feb',
-  'Mar',
-  'Apr',
-  'May',
-  'Jun',
-  'Jul',
-  'Aug',
-  'Sep',
-  'Oct',
-  'Nov',
-  'Dec',
-];
-
-String formatDate(DateTime dt) {
-  final weekday = _weekdays[dt.weekday - 1];
-  final month = _months[dt.month - 1];
-  return '$weekday, $month ${dt.day}';
-}
-
-String formatTime(DateTime dt) {
-  final hour = dt.hour % 12 == 0 ? 12 : dt.hour % 12;
-  final minute = dt.minute.toString().padLeft(2, '0');
-  final period = dt.hour < 12 ? 'AM' : 'PM';
-  return '$hour:$minute $period';
-}
-
-String formatDuration(Duration d) {
-  final hours = d.inHours;
-  final minutes = d.inMinutes.remainder(60);
-  if (hours > 0) return '${hours}h ${minutes}m';
-  return '${minutes}m';
-}
