+++ client/lib/utils/formatting.dart
@@ -1,0 +1,65 @@
+const weekdayAbbr = ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'];
+const weekdayFull = [
+  'Monday',
+  'Tuesday',
+  'Wednesday',
+  'Thursday',
+  'Friday',
+  'Saturday',
+  'Sunday',
+];
+const monthAbbr = [
+  'Jan',
+  'Feb',
+  'Mar',
+  'Apr',
+  'May',
+  'Jun',
+  'Jul',
+  'Aug',
+  'Sep',
+  'Oct',
+  'Nov',
+  'Dec',
+];
+
+/// Formats a DateTime as "Wed, Jan 5".
+String formatDate(DateTime dt) {
+  final weekday = weekdayAbbr[dt.weekday - 1];
+  final month = monthAbbr[dt.month - 1];
+  return '$weekday, $month ${dt.day}';
+}
+
+/// Formats a DateTime as "3:05 PM".
+String formatTime(DateTime dt) {
+  final hour = dt.hour % 12 == 0 ? 12 : dt.hour % 12;
+  final minute = dt.minute.toString().padLeft(2, '0');
+  final period = dt.hour < 12 ? 'AM' : 'PM';
+  return '$hour:$minute $period';
+}
+
+/// Formats a Duration as "2h 30m".
+String formatDuration(Duration d) {
+  final hours = d.inHours;
+  final minutes = d.inMinutes.remainder(60);
+  if (hours > 0) return '${hours}h ${minutes}m';
+  return '${minutes}m';
+}
+
+/// Formats hour and minute as "3:05 PM".
+String formatTimeOfDay(int hour, int minute) {
+  final h = hour % 12 == 0 ? 12 : hour % 12;
+  final m = minute.toString().padLeft(2, '0');
+  final period = hour < 12 ? 'AM' : 'PM';
+  return '$h:$m $period';
+}
+
+/// Formats seconds as "2h 30m".
+String formatSecsAsHoursMinutes(double secs) {
+  final totalMinutes = (secs / 60).round();
+  final hours = totalMinutes ~/ 60;
+  final minutes = totalMinutes % 60;
+  if (hours > 0) return '${hours}h ${minutes}m';
+  if (minutes > 0) return '${minutes}m';
+  return '0m';
+}
