+++ client/lib/views/kiosk/team_member_row.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:time_keeper/generated/common/common.pb.dart';
 import 'package:time_keeper/generated/db/db.pb.dart';
+import 'package:time_keeper/utils/formatting.dart';
 import 'package:time_keeper/utils/time.dart';
 
 class TeamMemberRow extends StatelessWidget {
@@ -18,11 +19,7 @@ class TeamMemberRow extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final alias = teamMember.alias.isNotEmpty ? '(${teamMember.alias})' : '';
-    final dt = timeIn.toDateTime();
-    final hour = dt.hour % 12 == 0 ? 12 : dt.hour % 12;
-    final minute = dt.minute.toString().padLeft(2, '0');
-    final period = dt.hour >= 12 ? 'PM' : 'AM';
-    final timeStr = '$hour:$minute $period';
+    final timeStr = formatTime(timeIn.toDateTime());
     return Row(
       children: [
         Expanded(child: Center(child: Text('${teamMember.firstName} $alias'))),
