+++ client/lib/views/sessions/session_stats.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:time_keeper/generated/db/db.pb.dart';
+import 'package:time_keeper/models/session_status.dart';
 import 'package:time_keeper/utils/time.dart';
-import 'package:time_keeper/views/sessions/session_helpers.dart';
+import 'package:time_keeper/widgets/stat_card.dart';
 
 class SessionStats extends StatelessWidget {
   final Map<String, Session> sessions;
@@ -35,93 +36,41 @@ class SessionStats extends StatelessWidget {
 
     return Row(
       children: [
-        _StatCard(
+        StatCard(
           icon: Icons.event,
           label: 'Total',
           value: '${sessions.length}',
           color: theme.colorScheme.primary,
         ),
         const SizedBox(width: 12),
-        _StatCard(
+        StatCard(
           icon: Icons.play_circle,
           label: 'Active',
           value: '$activeSessions',
           color: Colors.green,
         ),
         const SizedBox(width: 12),
-        _StatCard(
+        StatCard(
           icon: Icons.schedule,
           label: 'Upcoming',
           value: '$upcomingSessions',
           color: Colors.blue,
         ),
         const SizedBox(width: 12),
-        _StatCard(
+        StatCard(
           icon: Icons.calendar_month,
           label: 'This Month',
           value: '$thisMonth',
           color: theme.colorScheme.tertiary,
         ),
         const SizedBox(width: 12),
-        _StatCard(
+        StatCard(
           icon: Icons.people,
           label: 'Unique Members',
           value: '${uniqueMembers.length}',
           color: theme.colorScheme.secondary,
         ),
       ],
-    );
-  }
-}
-
-class _StatCard extends StatelessWidget {
-  final IconData icon;
-  final String label;
-  final String value;
-  final Color color;
-
-  const _StatCard({
-    required this.icon,
-    required this.label,
-    required this.value,
-    required this.color,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-      decoration: BoxDecoration(
-        color: color.withValues(alpha: 0.1),
-        borderRadius: BorderRadius.circular(8),
-        border: Border.all(color: color.withValues(alpha: 0.3)),
-      ),
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Icon(icon, color: color, size: 20),
-          const SizedBox(width: 8),
-          Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                value,
-                style: theme.textTheme.titleMedium?.copyWith(
-                  fontWeight: FontWeight.bold,
-                  color: color,
-                ),
-              ),
-              Text(
-                label,
-                style: theme.textTheme.bodySmall?.copyWith(
-                  color: theme.colorScheme.onSurfaceVariant,
-                ),
-              ),
-            ],
-          ),
-        ],
-      ),
     );
   }
 }
