+++ client/lib/views/stats/stats_overview_cards.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:time_keeper/generated/db/db.pb.dart';
 import 'package:time_keeper/utils/time.dart';
 import 'package:time_keeper/views/stats/stats_helpers.dart';
+import 'package:time_keeper/widgets/stat_card.dart';
 
 class StatsOverviewCards extends StatelessWidget {
   final Map<String, Session> filteredSessions;
@@ -46,95 +47,43 @@ class StatsOverviewCards extends StatelessWidget {
       spacing: 12,
       runSpacing: 12,
       children: [
-        _StatCard(
+        StatCard(
           icon: Icons.event,
           label: 'Total Sessions',
           value: '${filteredSessions.length}',
           color: theme.colorScheme.primary,
         ),
-        _StatCard(
+        StatCard(
           icon: Icons.schedule,
           label: 'Total Hours',
           value: formatSecsAsHoursMinutes(totalRegular + totalOvertime),
           color: theme.colorScheme.secondary,
         ),
-        _StatCard(
+        StatCard(
           icon: Icons.timer,
           label: 'Avg Session Duration',
           value: formatSecsAsHoursMinutes(avgSessionDuration),
           color: Colors.blue,
         ),
-        _StatCard(
+        StatCard(
           icon: Icons.warning_amber,
           label: 'Total Overtime',
           value: formatSecsAsHoursMinutes(totalOvertime),
           color: theme.colorScheme.error,
         ),
-        _StatCard(
+        StatCard(
           icon: Icons.people,
           label: 'Unique Members',
           value: '${insights.uniqueMembers}',
           color: Colors.green,
         ),
-        _StatCard(
+        StatCard(
           icon: Icons.groups,
           label: 'Avg Attendance',
           value: insights.avgAttendancePerSession.toStringAsFixed(1),
           color: theme.colorScheme.tertiary,
         ),
       ],
-    );
-  }
-}
-
-class _StatCard extends StatelessWidget {
-  final IconData icon;
-  final String label;
-  final String value;
-  final Color color;
-
-  const _StatCard({
-    required this.icon,
-    required this.label,
-    required this.value,
-    required this.color,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-      decoration: BoxDecoration(
-        color: color.withValues(alpha: 0.1),
-        borderRadius: BorderRadius.circular(8),
-        border: Border.all(color: color.withValues(alpha: 0.3)),
-      ),
-      child: Row(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Icon(icon, color: color, size: 20),
-          const SizedBox(width: 8),
-          Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                value,
-                style: theme.textTheme.titleMedium?.copyWith(
-                  fontWeight: FontWeight.bold,
-                  color: color,
-                ),
-              ),
-              Text(
-                label,
-                style: theme.textTheme.bodySmall?.copyWith(
-                  color: theme.colorScheme.onSurfaceVariant,
-                ),
-              ),
-            ],
-          ),
-        ],
-      ),
     );
   }
 }
