+++ client/lib/widgets/tables/header_text.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+class TableHeaderText extends StatelessWidget {
+  final String text;
+  const TableHeaderText(this.text, {super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Text(
+      text,
+      style: const TextStyle(
+        fontWeight: FontWeight.w600,
+        color: Colors.white,
+        fontSize: 13,
+      ),
+    );
+  }
+}
